MAJOR parameter overhaul. First added "getEditor" class to ParameterAPI which removes search paths for parameter editor creation. Then, added generics to ParameterAPIs and implemented generics for a few parameters (such as boolean, and double parameters). We still need to update the rest of the parameters to use generics.

+++ org/opensha/sha/imr/attenRelImpl/AS_2008_AttenRel.java
@@ -479,8 +479,8 @@ NamedObjectAPI, ParameterChangeListener {
 	 */
 	public void setSite(Site site) throws ParameterException {
 
-		vs30Param.setValue(site.getParameter(this.VS30_NAME).getValue());
-		depthTo1pt0kmPerSecParam.setValueIgnoreWarning(site.getParameter(this.DEPTH_1pt0_NAME).
+		vs30Param.setValue((Double)site.getParameter(this.VS30_NAME).getValue());
+		depthTo1pt0kmPerSecParam.setValueIgnoreWarning((Double)site.getParameter(this.DEPTH_1pt0_NAME).
 				getValue());
 		vs30_TypeParam.setValue(site.getParameter(VS30_TYPE_NAME).getValue());
