+++ scratchJavaDevelopers/jennie/attenRelImpl/CBJ_2008_AttenRel.java
@@ -466,8 +466,8 @@ public class CBJ_2008_AttenRel
    */
   public void setSite(Site site) throws ParameterException {
 
-    vs30Param.setValue(site.getParameter(this.VS30_NAME).getValue());
-    depthTo2pt5kmPerSecParam.setValueIgnoreWarning(site.getParameter(this.DEPTH_2pt5_NAME).
+    vs30Param.setValue((Double)site.getParameter(this.VS30_NAME).getValue());
+    depthTo2pt5kmPerSecParam.setValueIgnoreWarning((Double)site.getParameter(this.DEPTH_2pt5_NAME).
                                       getValue());
     this.site = site;
     setPropagationEffectParams();
